Create client-specific data.
This is a substantial refactor to allow for client-specific data, such
as buffers, request, and response. The context data structure was
renamed to represent the server data.

I also included uthash (http://troydhanson.github.io/uthash/) to allow
for quick access to the client data using the socket descriptor as the
key.